move RouterService to examples
