+++ crates/mcp-server/src/router.rs
@@ -1,7 +1,6 @@
 use std::{
     future::Future,
     pin::Pin,
-    task::{Context, Poll},
 };
 
 type PromptFuture = Pin<Box<dyn Future<Output = Result<String, PromptError>> + Send + 'static>>;
@@ -19,9 +18,8 @@ use mcp_core::{
     ResourceContents,
 };
 use serde_json::Value;
-use tower_service::Service;
 
-use crate::{BoxError, RouterError};
+use crate::RouterError;
 
 /// Builder for configuring and constructing capabilities
 pub struct CapabilitiesBuilder {
@@ -389,43 +387,5 @@ pub trait Router: Send + Sync + 'static {
             );
             Ok(response)
         }
-    }
-}
-
-pub struct RouterService<T>(pub T);
-
-impl<T> Service<JsonRpcRequest> for RouterService<T>
-where
-    T: Router + Clone + Send + Sync + 'static,
-{
-    type Response = JsonRpcResponse;
-    type Error = BoxError;
-    type Future = Pin<Box<dyn Future<Output = Result<Self::Response, Self::Error>> + Send>>;
-
-    fn poll_ready(&mut self, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        Poll::Ready(Ok(()))
-    }
-
-    fn call(&mut self, req: JsonRpcRequest) -> Self::Future {
-        let this = self.0.clone();
-
-        Box::pin(async move {
-            let result = match req.method.as_str() {
-                "initialize" => this.handle_initialize(req).await,
-                "tools/list" => this.handle_tools_list(req).await,
-                "tools/call" => this.handle_tools_call(req).await,
-                "resources/list" => this.handle_resources_list(req).await,
-                "resources/read" => this.handle_resources_read(req).await,
-                "prompts/list" => this.handle_prompts_list(req).await,
-                "prompts/get" => this.handle_prompts_get(req).await,
-                _ => {
-                    let mut response = this.create_response(req.id);
-                    response.error = Some(RouterError::MethodNotFound(req.method).into());
-                    Ok(response)
-                }
-            };
-
-            result.map_err(BoxError::from)
-        })
     }
 }
