+++ examples/servers/src/common/counter.rs
@@ -1,5 +1,4 @@
-use std::{future::Future, pin::Pin, sync::Arc};
-
+use mcp_core::protocol::{JsonRpcRequest, JsonRpcResponse};
 use mcp_core::{
     handler::{PromptError, ResourceError},
     prompt::{Prompt, PromptArgument},
@@ -7,8 +6,13 @@ use mcp_core::{
     Content, Resource, Tool, ToolError,
 };
 use mcp_server::router::CapabilitiesBuilder;
+use mcp_server::{BoxError, Router, RouterError};
 use serde_json::Value;
+use std::task::{Context, Poll};
+use std::{future::Future, pin::Pin, sync::Arc};
 use tokio::sync::Mutex;
+use tower_service::Service;
+
 
 #[derive(Clone)]
 pub struct CounterRouter {
@@ -179,6 +183,44 @@ impl mcp_server::Router for CounterRouter {
                     prompt_name
                 ))),
             }
+        })
+    }
+}
+
+pub struct RouterService<T>(pub T);
+
+impl<T> Service<JsonRpcRequest> for RouterService<T>
+where
+    T: Router + Clone + Send + Sync + 'static,
+{
+    type Response = JsonRpcResponse;
+    type Error = BoxError;
+    type Future = Pin<Box<dyn Future<Output =std::result::Result<Self::Response, Self::Error>> + Send>>;
+
+    fn poll_ready(&mut self, _cx: &mut Context<'_>) -> Poll<std::result::Result<(), Self::Error>> {
+        Poll::Ready(Ok(()))
+    }
+
+    fn call(&mut self, req: JsonRpcRequest) -> Self::Future {
+        let this = self.0.clone();
+
+        Box::pin(async move {
+            let result = match req.method.as_str() {
+                "initialize" => this.handle_initialize(req).await,
+                "tools/list" => this.handle_tools_list(req).await,
+                "tools/call" => this.handle_tools_call(req).await,
+                "resources/list" => this.handle_resources_list(req).await,
+                "resources/read" => this.handle_resources_read(req).await,
+                "prompts/list" => this.handle_prompts_list(req).await,
+                "prompts/get" => this.handle_prompts_get(req).await,
+                _ => {
+                    let mut response = this.create_response(req.id);
+                    response.error = Some(RouterError::MethodNotFound(req.method).into());
+                    Ok(response)
+                }
+            };
+
+            result.map_err(BoxError::from)
         })
     }
 }
