+++ examples/servers/src/counter_server.rs
@@ -1,5 +1,5 @@
 use anyhow::Result;
-use mcp_server::router::RouterService;
+use common::counter::RouterService;
 use mcp_server::{ByteTransport, Server};
 use tokio::io::{stdin, stdout};
 use tracing_appender::rolling::{RollingFileAppender, Rotation};
